Add files via upload
The url link of the webpage where university transcript is present is encoded. The above code encrypts the url link and provides a unique captcha which is to be used in the decryption file. The code is implemented for a single pc thus tcp server-client part contains a loopback address. The tcp-server client part is added to securely pass the captcha generated to the decryption code. Thus to run this code,run the above code and the decryption code at the same time in different command prompts.